Fixed hit Particle direction with bullets
Now when the player's ability bullet hits the boss, the impact effects go in the proper direction. In the Ability script of the specific abilities "FastProjectileAbility", or "SpreadShotAbility", I set the bullet's transform.up to the direction it is travelling at. So now when the boss takes damage and the transform.up is used, it will correctly get the opposite direction for the hit effect.

I also fixed another bug with the Painkiller's paracetamania when sucking in the pills there would be an error. I just made sure that the targetThrown != null since that was the cause of it.