Put all Events into one Scriptable Object
We had many different Scriptable Objects that just held events for different purposes. For each object that referenced a scriptable Object we had to make a SerializeField and attach it or else the game would throw an error.

Now I made 1 scriptable Object and put it in the resources folder that hosts ALL of the events we want to send. Now all Objects can reference the GameManager if they want to either receive or send an event.