Rehauled Boss Intro Animation
Now the boss intro animation script is removed and all the logic is placed in the InitializeBoss.cs script. This way we can access everything related to the boss on the boss game object and not have to scavenge the hierarchy.

I also made it so that if we hide the boss Intro Canvas game object, we will jump straight into gameplay instead of having to wait for a certain amount of time