Added Aiming Arrows Back
Now we've got the aim arrows back in the game and I made the hierarchy easier to manage by getting rid of the Aim Arrows Game Object and just making it a child in the Carry Object game object. There's also less calls to show and hide the aim arrows game object since I call it once when we initialize a carry object and once again when we press the launch projectile button.