Rehaul Applied Changes to All Scenes
Now each boss is playable with all the new changes made! There's still a few minor changes to be made with the projectiles, but other than that it all works fine.

I made duplicate scenes for the Grape and Dairy bosses with the new changes, the other ones are obsolete.