Proj Pattern: Fixed Clear All Button and Angles
The clear all button stopped working because clicking anywhere in the inspector to deselect the currently selected pattern was eating up the input. All I had to do was place that code underneath the button code for it to work again.

Angles weren't working either. Only the original projectile was being sent to the correct angle but the others just to the right. I added a calculation to get the angle at the start of the method and applied it to everytime we need to get the angle to work.