Proj Pattern: Refreshing UI Contents on Code Change
Restructured the code base and how we initialize the pattern modifiers, so now when users make a change in the code by editing the modName, value, variableType, tooltip, etc. It will automatically refresh the inspector to match it.
Before this change users had to delete the SO and make a new one. Good riddance to that! Users no longer have to delete our hard work : )

Added the ability to write tooltips for each pattern modifier! Now when the user hovers over the UI it will display the tooltip giving them the necessary information of what the modifier does :)

Added "Lifetime duration" in the basePAT but have not yet implemented its functionality with the Spawner.
The BasePAT's modNames also changed to include the word "Initial" before the words "angle, speed, delay, etc"

Organized the folder structure a bit more to isolate scripts and prefabs into their own folders.

Deleted all pre-existing Projectile Pattern Scriptable Objects since they no longer worked with the new updates.